api/Python/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.4.0
async-timeout==4.0.2
attrs==22.2.0
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
#azure-search-documents==11.4.0b4
./azure_search_documents-11.4.0b4-py3-none-any.whl
azure-storage-blob==12.16.0
backoff==2.2.1
beautifulsoup4==4.11.2
blobfile==2.0.1
boto3==1.26.108
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==39.0.2
cssutils==1.0.2
dataclasses-json==0.5.7
Deprecated==1.2.13
dnspython==2.3.0
et-xmlfile==1.1.0
fake-useragent==1.1.1
filelock==3.9.0
frozenlist==1.3.3
greenlet==2.0.2
grpcio==1.47.5
grpcio-tools==1.47.5
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
joblib==1.2.0
langchain==0.0.177
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
mmh3==3.0.0
monotonic==1.6
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.6
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pathos==0.2.9
pdfminer.six==20221105
Pillow==9.4.0
pinecone-client==2.2.1
portalocker==2.7.0
protobuf==3.20.3
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.6
Pygments==2.14.0
PyJWT==2.6.0
pymilvus==2.2.2
pyodbc==4.0.35
pymupdf==1.21.1
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
tenacity==8.2.2
tiktoken==0.3.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
ujson==5.4.0
unstructured==0.5.2
urllib3==1.26.15
weaviate-client==3.15.2
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
zipp==3.15.0
